+++ builtins/env.c
@@ -1,22 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exit.c                                             :+:      :+:    :+:   */
+/*   env.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: asoler <asoler@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/08 21:35:14 by asoler            #+#    #+#             */
-/*   Updated: 2022/12/05 21:24:22 by asoler           ###   ########.fr       */
+/*   Created: 2022/12/05 21:29:01 by asoler            #+#    #+#             */
+/*   Updated: 2022/12/05 21:29:30 by asoler           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-void	builtin_exit(t_data *data)
+void	builtin_env(t_env *hash_table[TABLE_SIZE])
 {
-	free_and_count_array(data->path, free);
-	free_hash_table(data);
-	ft_cmdclear(&data->cmds, free);
-	free(data->line);
-	exit(0);
+	int	i;
+
+	i = 0;
+	while (i < TABLE_SIZE)
+	{
+		if (hash_table[i])
+			print_env_lst(hash_table[i]);
+		i++;
+	}
 }
